requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
scikit-learn==1.5.2
numpy==2.1.1
pandas==2.2.3
python-dotenv==1.0.1
loguru==0.7.2
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
prometheus-client==0.21.0
shap==0.45.1

